refactor: LineGenerator이 생성하는 객체의 타입 변경
LineGenerator가 Points 객체를 생성하도록 변경
