+++ src/main/java/domain/DefaultLineGenerator.java
@@ -7,17 +7,17 @@ import java.util.Random;
 public class DefaultLineGenerator implements LineGenerator {
 
     @Override
-    public Line generateLine(int personNumber) {
+    public Points generateLine(int personNumber) {
         int pointsNumber = personNumber - 1;
         boolean beforeValue = false;
-        List<Boolean> points = new ArrayList<>();
+        List<Point> points = new ArrayList<>();
         for (int i = 0; i < pointsNumber; i++) {
             boolean nextValue = generateNextValue(beforeValue);
-            points.add(nextValue);
+            points.add(new Point(nextValue));
             beforeValue = nextValue;
         }
 
-        return new Line(points);
+        return new Points(points);
     }
 
     private boolean generateNextValue(boolean beforeValue) {
