+++ src/main/java/domain/Ladder.java
@@ -1,5 +1,9 @@
 package domain;
 
+import dto.ResultRequest;
+import dto.ResultsRequest;
+import dto.ResultsResponse;
+
 import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
