+++ src/main/java/domain/Line.java
@@ -1,16 +1,13 @@
 package domain;
 
-import java.util.List;
-import java.util.stream.Collectors;
-
 public class Line {
     private final Points points;
 
-    protected Line(List<Boolean> points) {
-        this.points = new Points(points.stream()
-                .map(Point::new)
-                .collect(Collectors.toList()));
-    }
+//    protected Line(List<Boolean> points) {
+//        this.points = new Points(points.stream()
+//                .map(Point::new)
+//                .collect(Collectors.toList()));
+//    }
 
     public Line(Points points) {
         this.points = new Points(points.getValues());
