+++ src/test/java/dto/ResultsResponseTest.java
@@ -1,4 +1,4 @@
-package domain;
+package dto;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
